CIFAR10

+++ data_loader/data_utils.py
@@ -11,13 +11,11 @@ import json
 import wget
 from definitions import *
 
-
 # datasets path
 if not os.path.exists(os.path.abspath(DATASETS_DIR)):
     os.makedirs(os.path.abspath(DATASETS_DIR))
 if not os.path.exists(os.path.abspath(DATASETS_METADATA_DIR)):
     os.makedirs(os.path.abspath(DATASETS_METADATA_DIR))
-
 
 # Logging
 # logger = logging.getLogger(name=__name__)
@@ -27,6 +25,8 @@ if not os.path.exists(os.path.abspath(DATASETS_METADATA_DIR)):
 """
 get_<dataset_name>() functions below are used to download the corresponding dataset and save their metadata.
 """
+
+
 def get_creditcard():
     file_name = os.path.join(DATASETS_DIR, 'creditcard.csv')
     if not os.path.exists(os.path.join(DATASETS_DIR, 'creditcard.csv')):
@@ -45,7 +45,7 @@ def get_creditcard():
                  "model_args": {
                      "input_size": 28,
                      "output_size": 2
-                    }
+                 }
                  }
         label_values = data["Class"].unique()
         for label in label_values:
@@ -57,22 +57,22 @@ def get_creditcard():
 
 def get_mnist():
     train_set = torchvision.datasets.MNIST(
-            root=DATASETS_DIR, train=True, download=True,
-            transform=torchvision.transforms.Compose([
-                torchvision.transforms.ToTensor(),
-                # torchvision.transforms.Normalize((0.1307,), (0.3081,))
-            ])
-        )
+        root=DATASETS_DIR, train=True, download=True,
+        transform=torchvision.transforms.Compose([
+            torchvision.transforms.ToTensor(),
+            # torchvision.transforms.Normalize((0.1307,), (0.3081,))
+        ])
+    )
     test_set = torchvision.datasets.MNIST(
-            root=DATASETS_DIR, train=False, download=True,
-            transform=torchvision.transforms.Compose([
-                torchvision.transforms.ToTensor(),
-                # torchvision.transforms.Normalize((0.1307,), (0.3081,))
-            ])
-        )
+        root=DATASETS_DIR, train=False, download=True,
+        transform=torchvision.transforms.Compose([
+            torchvision.transforms.ToTensor(),
+            # torchvision.transforms.Normalize((0.1307,), (0.3081,))
+        ])
+    )
     if not os.path.exists(os.path.join(DATASETS_METADATA_DIR, 'mnist.json')):
         metad = {"type": "torchvisionDataset",
-                 "num_features": 28*28,
+                 "num_features": 28 * 28,
                  "num_classes": 10,
                  "classes": list(range(0, 10)),
                  "num_samples": {
@@ -84,15 +84,59 @@ def get_mnist():
                      "test": {i: 0 for i in range(0, 10)},
                  },
                  "model_args": {
-                     "input_size": 28*28,
-                     "output_size": 10,
-                    }
+                     "input_size": 1,
+                     "output_size": 2,
+                     "input_img_size": (28, 28)
+                 }
                  }
         for sample, label in train_set:
             metad["num_samples_per_class"]["train"][label] += 1
         for sample, label in test_set:
             metad["num_samples_per_class"]["test"][label] += 1
         with open(os.path.join(DATASETS_METADATA_DIR, 'mnist.json'), 'w') as f:
+            json.dump(metad, f, indent=4)
+    return
+
+
+def get_cifar10():
+    train_set = torchvision.datasets.CIFAR10(
+        root=DATASETS_DIR, train=True, download=True,
+        transform=torchvision.transforms.Compose([
+            torchvision.transforms.ToTensor(),
+            # torchvision.transforms.Normalize((0.1307,), (0.3081,))
+        ])
+    )
+    test_set = torchvision.datasets.CIFAR10(
+        root=DATASETS_DIR, train=False, download=True,
+        transform=torchvision.transforms.Compose([
+            torchvision.transforms.ToTensor(),
+            # torchvision.transforms.Normalize((0.1307,), (0.3081,))
+        ])
+    )
+    if not os.path.exists(os.path.join(DATASETS_METADATA_DIR, 'mnist.json')):
+        metad = {"type": "torchvisionDataset",
+                 "num_features": 32 * 32,
+                 "num_classes": 10,
+                 "classes": list(range(0, 10)),
+                 "num_samples": {
+                     "train": len(train_set),
+                     "test": len(test_set)
+                 },
+                 "num_samples_per_class": {
+                     "train": {i: 0 for i in range(0, 10)},
+                     "test": {i: 0 for i in range(0, 10)},
+                 },
+                 "model_args": {
+                     "input_size": 1,
+                     "output_size": 2,
+                     "input_img_size": (32, 32)
+                 }
+                 }
+        for sample, label in train_set:
+            metad["num_samples_per_class"]["train"][label] += 1
+        for sample, label in test_set:
+            metad["num_samples_per_class"]["test"][label] += 1
+        with open(os.path.join(DATASETS_METADATA_DIR, 'cifar10.json'), 'w') as f:
             json.dump(metad, f, indent=4)
     return
